+++ network_toolkit/inventory/__init__.py
@@ -1,4 +1,5 @@
 from .csv_import import import_switches_from_csv
+from .prime_import import import_switches_from_prime
 # from .prime_import import *
 
-__all__ = ["import_switches_from_csv"]
+__all__ = ["import_switches_from_csv", "import_switches_from_prime"]